+++ pkg/contrib/main.ts
@@ -30,4 +30,10 @@ export * from "./src/systems/ControllerSystem";
 export * from "./src/systems/ClickableSystem";
 export * from "./src/systems/NetcodeSystem";
 export * from "./src/systems/PlayerSpawnSystem";
-export * from "./src/entities/Entities";
+export * from "./src/entities/Ball";
+export * from "./src/entities/DebugButton";
+export * from "./src/entities/FpsText";
+export * from "./src/entities/FullscreenButton";
+export * from "./src/entities/Skelly";
+export * from "./src/entities/Spaceship";
+export * from "./src/entities/TileFloor";
